+++ packages/genesys-spark-components/src/components/stable/gux-form-field/functional-components/gux-form-field-error/gux-form-field-error.tsx
@@ -15,7 +15,11 @@ export const GuxFormFieldError: FunctionalComponent<GuxFormFieldErrorProps> = (
         'gux-show': show
       }}
     >
-      <gux-icon icon-name="fa/hexagon-exclamation-solid" decorative></gux-icon>
+      <gux-icon
+        icon-name="fa/hexagon-exclamation-solid"
+        decorative
+        size="small"
+      ></gux-icon>
       <div class="gux-message">{children}</div>
     </div>
   ) as VNode;