+++ packages/genesys-spark-components/src/components/stable/gux-table/gux-table-select-menu/gux-table-select-menu.tsx
@@ -120,6 +120,7 @@ export class GuxTableSelectMenu {
           <gux-icon
             icon-name="custom/chevron-down-small-regular"
             screenreader-text={this.i18n('tableOptions')}
+            size="small"
           ></gux-icon>
         </button>,
         <gux-popover-list